feat: Handle exception when creating user in SignUpForm
This commit handles the exception that may occur when creating a user in the `SignUpForm` class. If an exception is raised during the creation of the `UserCreate` object, it will be caught and re-raised to provide more detailed information about the error. This ensures that any errors during user creation are properly handled and can be communicated to the client.